Refactor script into package modules
