--- MIDI-mapping.md
@@ -1,97 +1,0 @@
-# Mapping of General MIDI Percussion Key Map (GMPKM) to three frequency levels
-
-## Note	-	Name	-	Category
-
-35	-	Acoustic Bass Drum	-	low
-
-36	-	Bass Drum 1	-	low
-
-37	-	Side Stick	-	mid
-
-38	-	Acoustic Snare	-	mid
-
-39	-	Hand Clap	-	mid
-
-40	-	Electric Snare	-	mid
-
-41	-	Low Floor Tom	-	low
-
-42	-	Closed Hi Hat	-	high
-
-43	-	High Floor Tom	-	mid
-
-44	-	Pedal Hi-Hat	-	high
-
-45	-	Low Tom	-	low
-
-46	-	Open Hi-Hat	-	high
-
-47	-	Low-Mid Tom	-	mid
-
-48	-	Hi-Mid Tom	-	mid
-
-49	-	Crash Cymbal 1	-	high
-
-50	-	High Tom	-	mid
-
-51	-	Ride Cymbal 1	-	high
-
-52	-	Chinese Cymbal	-	high
-
-53	-	Ride Bell	-	high
-
-54	-	Tambourine	-	high
-
-55	-	Splash Cymbal	-	high
-
-56	-	Cowbell	-	high
-
-57	-	Crash Cymbal 2	-	high
-
-58	-	Vibraslap	-	mid
-
-59	-	Ride Cymbal 2	-	high
-
-60	-	Hi Bongo	-	high
-
-61	-	Low Bongo	-	mid
-
-62	-	Mute Hi Conga	-	high
-
-63	-	Open Hi Conga	-	high
-
-64	-	Low Conga	-	low
-
-65	-	High Timbale	-	mid
-
-66	-	Low Timbale	-	low
-
-67	-	High Agogo	-	high
-
-68	-	Low Agogo	-	mid
-
-69	-	Cabasa	-	high
-
-70	-	Maracas	-	high
-
-71	-	Short Whistle	-	high
-
-72	-	Long Whistle	-	high
-
-73	-	Short Guiro	-	high
-
-74	-	Long Guiro	-	high
-
-75	-	Claves	-	high
-
-76	-	Hi Wood Block	-	high
-
-77	-	Low Wood Block	-	low
-
-78	-	Mute Cuica	-	high
-
-79	-	Open Cuica	-	high
-
-80	-	Mute Triangle	-	high
-
-81	-	Open Triangle	-	high
